+++ src/animals/Cat.java
@@ -1,16 +1,9 @@
 package animals;
 
-// Subclass Cat that inherits from Animal
+// Cat class that extends Animal
 class Cat extends Animal {
-    // Constructor
-    public Cat(String name, int age) {
-        super(name, age);
-    }
-
-    // Implementing abstract method from Animal class
-    @Override
-    public void makeSound() {
-        System.out.println(getName() + " says: Meow!");
+    public Cat(String name) {
+        super(name);
     }
 }
