+++ src/animals/Main.java
@@ -1,24 +1,16 @@
 package animals;
 
-// Testing the classes
+// Main class to demonstrate the functionality
 public class Main {
     public static void main(String[] args) {
-        // Cannot create an instance of the abstract class Animal
-        // Animal animal = new Animal("Generic Animal", 5); // This will cause an error!
+        // Create objects of Dog, Cat, and Lion classes
+        Dog dog = new Dog("Tarzan");
+        Cat cat = new Cat("Garfield");
+        Lion lion = new Lion("Simba");
 
-        // Create instances of the subclasses
-        Dog dog = new Dog("Buddy", 3);
-        Cat cat = new Cat("Whiskers", 2);
-        Bird bird = new Bird("Tweety", 1);
-
-        // Calling methods on the objects
-        dog.eat();   // Buddy is eating.
-        dog.makeSound();  // Buddy says: Woof Woof!
-
-        cat.eat();   // Whiskers is eating.
-        cat.makeSound();  // Whiskers says: Meow!
-
-        bird.eat();   // Tweety is eating.
-        bird.makeSound();  // Tweety says: Tweet!
+        // Use action method
+        dog.action();  // Output: Dog is eating
+        cat.action();  // Output: Cat is eating
+        lion.action(); // Output: Lion is eating
     }
 }
