--- src/animalsOfPrince/Cat.java
@@ -1,9 +1,0 @@
-package animalsOfPrince;
-
-// Cat class that extends Animal
-class Cat extends Animal {
-    public Cat(String name) {
-        super(name);
-    }
-}
-
