--- src/animalsOfPrince/Dog.java
@@ -1,9 +1,0 @@
-package animalsOfPrince;
-
-// Dog class that extends Animal
-class Dog extends Animal {
-    public Dog(String name) {
-        super(name);
-    }
-}
-
